Profile performance improvements
